opt/cr3di/requirements.txt
typer[all]
pydantic
python-gnupg
InquirerPy
